Port to Qt 4.6 etc...
Ported to use Qt 4.6's matrix and vector classes, as well as using
QGLShaderProgram to make the OpenGL rendering code work both on desktop
OpenGL and OpenGL ES 2.0 (the desktop version requires GLEW atm).

Also added auto-walk toggler to improve navigation on a touchscreen
device.

+++ entity.cpp
@@ -122,9 +122,9 @@ bool Entity::move(MazeScene *scene)
                 angleToTarget -= 360;
 
             if (angleToTarget < 0)
-                m_angle -= qMin(-angleToTarget, 0.5);
+                m_angle -= qMin(-angleToTarget, qreal(0.5));
             else
-                m_angle += qMin(angleToTarget, 0.5);
+                m_angle += qMin(angleToTarget, qreal(0.5));
             moved = true;
         }
     } else if (m_turnVelocity != 0) {
